feat: EYE-49 Study Statistics, Mock separation
